requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
joblib>=1.1.0
matplotlib>=3.5.0
openaq>=3.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.24.0
pydantic>=2.0.0
